feat: move routes to app router
