+++ app/app/api/(ai)/vision/route.ts
@@ -1,11 +1,6 @@
-import type { NextApiRequest, NextApiResponse } from "next";
-
-type ResponseData = {
-  message: string;
-};
+import { NextResponse } from "next/server";
 
 export const config = {
-  // increase max size
   api: {
     bodyParser: {
       sizeLimit: "20mb",
@@ -13,13 +8,10 @@ export const config = {
   },
 };
 
-export default async function handler(
-  req: NextApiRequest,
-  res: NextApiResponse<ResponseData>
-) {
+export async function POST(request: Request) {
   try {
     const apiKey = process.env.OPENAI_API_KEY || "";
-    const payload = req.body;
+    const payload = await request.json();
     console.log("payload text route", payload);
 
     const model = "gpt-4-turbo";
@@ -40,19 +32,20 @@ export default async function handler(
 
     if (!response.ok) {
       console.log(`Error: ${response.status}`);
-      return res.status(response.status).json({
-        message: `Server responded with status: ${response.status}`,
-      });
+      return NextResponse.json(
+        { message: `Server responded with status: ${response.status}` },
+        { status: response.status }
+      );
     }
 
     if (response.status === 401) {
       console.log("Invalid API key");
-      return res.status(401).json({ message: "Invalid API key" });
+      return NextResponse.json({ message: "Invalid API key" }, { status: 401 });
     }
     const result = await response.json();
-    res.status(200).json(result);
+    return NextResponse.json(result);
   } catch (error) {
     console.error(error);
-    res.status(500).json({ message: "Error" });
+    return NextResponse.json({ message: "Error" }, { status: 500 });
   }
 }
